// src/interfaces/IPost.tsx
export type IPostData = {
  id?: string;
  title?: string | undefined;
  subtitle?: string | undefined;
  date?: string | undefined;
  type?: string[] | undefined;
  contentHtml?: string | undefined;
  description?: string | undefined;
  featuredImage?: string | undefined;
  category?: string[] | undefined;
};